Rename Sensor_Electronic dir, change import path Signed-off-by: Jun Jiang <jun.jiang@eonerc.rwth-aachen.de>
